Renamed useFileContent > useFileToText
